Correction du type d'identifiant dans DPIRepository pour utiliser Long au lieu de Integer et mise à jour des imports de styles dans plusieurs composants pour utiliser Table.css.

+++ back/src/main/java/esiag/back/models/medical/DPI.java
@@ -12,7 +12,7 @@ public class DPI {
     @Id
     @Column(name="id_dpi")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idDPI;
+    private Long idDPI;
 
     @ManyToOne
     @JoinColumn(name = "id_patient")
